refactor and make subpackages
