+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/FieldPosition.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.common;
 
 public enum FieldPosition {
     UNKNOWN,
