+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/ScoringElementLocation.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.common;
 
 public enum ScoringElementLocation {
     UNKNOWN,
