+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/processors/FirstVisionProcessor.java
@@ -4,8 +4,8 @@ import android.graphics.Canvas;
 
 
 import org.firstinspires.ftc.robotcore.internal.camera.calibration.CameraCalibration;
-import org.firstinspires.ftc.teamcode.Alliance;
-import org.firstinspires.ftc.teamcode.ScoringElementLocation;
+import org.firstinspires.ftc.teamcode.common.Alliance;
+import org.firstinspires.ftc.teamcode.common.ScoringElementLocation;
 import org.firstinspires.ftc.vision.VisionProcessor;
 import org.opencv.core.Core;
 import org.opencv.core.CvType;
